Merged dev into master, prepare release 2.1.0

+++ PerimeterXModule/DataContracts/Cookies/DecodedCookieV3.cs
@@ -11,7 +11,7 @@ namespace PerimeterX
 		[DataMember(Name = "t")]
 		public double Time { get; set; }
 		[DataMember(Name = "s")]
-		public double Score;
+		public int Score;
 		[DataMember(Name = "a")]
 		public string Action;
 	}
